Restructure data source files
