+++ src/calliope/preprocess/model_data.py
@@ -4,8 +4,7 @@
 import itertools
 import logging
 from copy import deepcopy
-from pathlib import Path
-from typing import Hashable, Literal, Optional
+from typing import Literal, Optional
 
 import numpy as np
 import pandas as pd
@@ -17,27 +16,11 @@ from calliope import exceptions
 from calliope.attrdict import AttrDict
 from calliope.preprocess import time
 from calliope.util.schema import MODEL_SCHEMA, validate_dict
-from calliope.util.tools import listify, relative_path
+from calliope.util.tools import listify
 
 LOGGER = logging.getLogger(__name__)
 
 DATA_T = Optional[float | int | bool | str] | list[Optional[float | int | bool | str]]
-
-PROTECTED_PARAMS = {
-    "to": "`to`/`from` is only valid in YAML definition. Set the `carrier_in` and `carrier_out` parameters of transmission technologies at the nodes they link together to emulate their function when loading from file.",
-    "from": "`to`/`from` is only valid in YAML definition. Set the `carrier_in` and `carrier_out` parameters of transmission technologies at the nodes they link together to emulate their function when loading from file.",
-    "definition_matrix": "`definition_matrix` is a protected array, it will be generated internally based on the values you assign to the `carrier_in` and `carrier_out` parameters.",
-    "inherit": "Cannot use technology/node inheritance (`inherit`) when loading data from file.",
-}
-
-
-class DataSource(TypedDict):
-    rows: NotRequired[str | list[str]]
-    columns: NotRequired[str | list[str]]
-    file: str
-    add_dimensions: NotRequired[dict[str, str | list[str]]]
-    drop: Hashable | list[Hashable]
-    sel_drop: dict[str, str | bool | int]
 
 
 class Param(TypedDict):
@@ -52,13 +35,94 @@ class ModelDefinition(TypedDict):
     tech_groups: NotRequired[AttrDict]
     node_groups: NotRequired[AttrDict]
     parameters: NotRequired[AttrDict]
-    data_sources: NotRequired[list[DataSource]]
 
 
 LOGGER = logging.getLogger(__name__)
 
 
 class ModelDataFactory:
+    def __init__(
+        self,
+        model_config: dict,
+        model_definition: ModelDefinition,
+        tech_data_from_sources: dict,
+    ):
+        """Take a Calliope model definition dictionary and convert it into an xarray Dataset, ready for
+        constraint generation.
+
+        This includes extracting timeseries data from file and resampling/clustering as necessary.
+
+        Args:
+            model_config (dict): Model initialisation configuration (i.e., `config.init`).
+            model_definition (ModelDefinition): Definition of model nodes and technologies as a dictionary.
+            attributes (dict): Attributes to attach to the model Dataset.
+            param_attributes (dict[str, dict]): Attributes to attach to the generated model DataArrays.
+        """
+
+        self.config: dict = model_config
+        self.model_definition: ModelDefinition = model_definition.copy()
+        self.tech_data_from_sources: AttrDict = AttrDict(tech_data_from_sources)
+
+    def climb_inheritance_tree(
+        self,
+        dim_item_dict: AttrDict,
+        dim_name: Literal["nodes", "techs"],
+        item_name: str,
+        inheritance: Optional[list] = None,
+    ) -> tuple[AttrDict, Optional[list]]:
+        """Follow the `inherit` references from `nodes` to `node_groups` / from `techs` to `tech_groups`.
+
+        Abstract group definitions (those in `node_groups`/`tech_groups`) can inherit each other, but `nodes`/`techs` cannot.
+
+        This function will be called recursively until a definition dictionary without `inherit` is reached.
+
+        Args:
+            dim_item_dict (AttrDict):
+                Dictionary (possibly) containing `inherit`. If it doesn't contain `inherit`, the climbing stops here.
+            dim_name (Literal[nodes, techs]):
+                The name of the dimension we're working with, so that we can access the correct `_groups` definitions.
+            item_name (str):
+                The current position in the inheritance tree.
+            inheritance (Optional[list], optional):
+                A list of items that have been inherited (starting with the oldest).
+                If the first `dim_item_dict` does not contain `inherit`, this will remain as None.
+                Defaults to None.
+
+        Raises:
+            KeyError: Must inherit from a named group item in `node_groups` (for `nodes`) and `tech_groups` (for `techs`)
+
+        Returns:
+            tuple[AttrDict, Optional[list]]: Definition dictionary with inherited data and a list of the inheritance tree climbed to get there.
+        """
+
+        to_inherit = dim_item_dict.get("inherit", None)
+        dim_groups = AttrDict(
+            self.model_definition.get(f"{dim_name.removesuffix('s')}_groups", {})
+        )
+        if to_inherit is None:
+            if dim_name == "techs" and item_name in self.tech_data_from_sources:
+                _data_source_dict = deepcopy(self.tech_data_from_sources[item_name])
+                _data_source_dict.union(dim_item_dict, allow_override=True)
+                dim_item_dict = _data_source_dict
+            updated_dim_item_dict = dim_item_dict
+        elif to_inherit not in dim_groups:
+            raise KeyError(
+                f"({dim_name}, {item_name}) | Cannot find `{to_inherit}` in inheritance tree."
+            )
+        else:
+            base_def_dict, inheritance = self.climb_inheritance_tree(
+                dim_groups[to_inherit], dim_name, to_inherit, inheritance
+            )
+            updated_dim_item_dict = deepcopy(base_def_dict)
+            updated_dim_item_dict.union(dim_item_dict, allow_override=True)
+            if inheritance is not None:
+                inheritance.append(to_inherit)
+            else:
+                inheritance = [to_inherit]
+        return updated_dim_item_dict, inheritance
+
+
+class ModelDefinitionToDataFactory(ModelDataFactory):
     # TODO: move into yaml syntax and have it be updatable
     LOOKUP_PARAMS = {
         "carrier_in": "carriers",
@@ -84,9 +148,10 @@ class ModelDataFactory:
         self,
         model_config: dict,
         model_definition: ModelDefinition,
+        initial_model_data: xr.Dataset,
+        tech_data_from_sources: dict,
         attributes: dict,
         param_attributes: dict[str, dict],
-        model_definition_path: Optional[Path] = None,
     ):
         """Take a Calliope model definition dictionary and convert it into an xarray Dataset, ready for
         constraint generation.
@@ -99,12 +164,8 @@ class ModelDataFactory:
             attributes (dict): Attributes to attach to the model Dataset.
             param_attributes (dict[str, dict]): Attributes to attach to the generated model DataArrays.
         """
-
-        self.config: dict = model_config
-        self.model_definition: ModelDefinition = model_definition.copy()
-        self.model_def_path = model_definition_path
-        self.model_data = xr.Dataset(attrs=AttrDict(attributes))
-        self._tech_data_source_dict = AttrDict()
+        super().__init__(model_config, model_definition, tech_data_from_sources)
+        self.model_data = initial_model_data.assign_attrs(AttrDict(attributes))
 
         flipped_attributes: dict[str, dict] = dict()
         for key, val in param_attributes.items():
@@ -122,7 +183,6 @@ class ModelDataFactory:
         Returns:
             xr.Dataset: Built model dataset, including the timeseries dimension.
         """
-        self.load_data_sources()
         self.add_node_tech_data()
         self.add_time_dimension(timeseries_dfs)
         self.add_top_level_params()
@@ -132,102 +192,6 @@ class ModelDataFactory:
         self.resample_time_dimension()
         self.assign_input_attr()
         return self.model_data
-
-    def load_data_sources(self) -> None:
-        """Load data from the `data_sources` list into the model dataset.
-
-        Each subsequent data source in the list will override prior data.
-
-        Raises:
-            exceptions.ModelError: All data sources must have a `parameters` dimension.
-            exceptions.ModelError: Certain parameters are only valid if defined in YAML; they cannot be defined in data sources.
-        """
-        datasets: list[xr.Dataset] = []
-        for data_source in self.model_definition.get("data_sources", []):
-            filename = data_source["file"]
-            columns = self._listify_if_defined(data_source, "columns")
-            index = self._listify_if_defined(data_source, "rows")
-
-            csv_reader_kwargs = {"header": columns, "index_col": index}
-            for axis, names in csv_reader_kwargs.items():
-                if names is not None:
-                    csv_reader_kwargs[axis] = [i for i, _ in enumerate(names)]
-
-            filepath = relative_path(self.model_def_path, filename)
-            df = pd.read_csv(filepath, encoding="utf-8", **csv_reader_kwargs)
-
-            for axis, names in {"columns": columns, "index": index}.items():
-                if names is None:
-                    df = df.squeeze(axis=axis)
-                else:
-                    self._compare_axis_names(
-                        getattr(df, axis).names, names, axis, filename
-                    )
-                    df.rename_axis(inplace=True, **{axis: names})
-
-            tdf: pd.Series
-            if isinstance(df, pd.DataFrame):
-                tdf = df.stack(df.columns.names)  # type: ignore
-            else:
-                tdf = df
-
-            if "drop" in data_source.keys():
-                tdf = tdf.droplevel(data_source["drop"])
-
-            if "sel_drop" in data_source.keys():
-                tdf = tdf.xs(
-                    tuple(data_source["sel_drop"].values()),
-                    level=tuple(data_source["sel_drop"].keys()),
-                    drop_level=True,
-                )
-
-            if "add_dimensions" in data_source.keys():
-                for dim_name, index_items in data_source["add_dimensions"].items():
-                    index_items = listify(index_items)
-                    tdf = pd.concat(
-                        [tdf for _ in index_items], keys=index_items, names=[dim_name]
-                    )
-
-            if "parameters" not in tdf.index.names:
-                raise exceptions.ModelError(
-                    f"(data_sources, {filename}) | The `parameters` dimension must exist in on of `rows`, `columns`, or `add_dimensions`."
-                )
-
-            invalid_params = tdf.index.get_level_values("parameters").intersection(
-                list(PROTECTED_PARAMS.keys())
-            )
-            if not invalid_params.empty:
-                extra_info = set(PROTECTED_PARAMS[k] for k in invalid_params)
-                exceptions.print_warnings_and_raise_errors(
-                    errors=list(extra_info), during=f"data source loading ({filename})"
-                )
-
-            if tdf.index.names == ["parameters"]:  # unindexed parameters
-                ds = xr.Dataset(tdf.to_dict())
-            else:
-                ds = tdf.unstack("parameters").to_xarray()
-            ds = time.clean_data_source_timeseries(ds, self.config, filename)
-            for da_name, da in ds.data_vars.items():
-                da.attrs.update(self.param_attrs.get(da_name, {}))
-            LOGGER.debug(f"(data_sources, {filename}) | Loaded arrays:\n{ds.data_vars}")
-            datasets.append(ds)
-
-        # We apply these 1. for each dataset separately and 2. for tech defs entirely before node defs.
-        # 1. because we want to avoid the affect of dimension broadcasting that will occur when merging the datasets.
-        # 2. because we want to ensure any possible `parent` info is available for techs when we update the node definitions.
-        for ds in datasets:
-            if "techs" in ds.dims:
-                self._update_tech_def_from_data_source(ds)
-        for ds in datasets:
-            if "techs" in ds.dims and "nodes" in ds.dims:
-                self._update_node_def_from_data_source(ds)
-
-        if datasets:
-            self.model_data = xr.merge(
-                [self.model_data, *datasets],
-                compat="override",
-                combine_attrs="no_conflicts",
-            )
 
     def add_node_tech_data(self):
         """For each node, extract technology definitions and node-level parameters and convert them to arrays.
@@ -466,77 +430,9 @@ class ModelDataFactory:
     def assign_input_attr(self):
         """All input parameters need to be assigned the `is_result=False` attribute to be able to filter the arrays in the calliope.Model object."""
         for var_name, var_data in self.model_data.data_vars.items():
-            self.model_data[var_name] = var_data.assign_attrs(is_result=False)
-
-    def _update_tech_def_from_data_source(self, data_source: xr.Dataset) -> None:
-        """Create a dummy technology definition dictionary from the dimensions defined across all data sources.
-
-        This definition dictionary will ensure that the minimal YAML content is still possible.
-
-        This function should be run _before_ `self._update_node_def_from_data_source`.
-
-        Args:
-            data_source (xr.Dataset): Data loaded from one file.
-        """
-        techs = data_source.techs
-        base_tech_data = AttrDict({k: {} for k in techs.values})
-        if "parent" in data_source:
-            parent_dict = AttrDict(data_source["parent"].to_dataframe().T.to_dict())
-            base_tech_data.union(parent_dict)
-        carrier_info_dict = self._carrier_info_dict_from_model_data_var(data_source)
-        base_tech_data.union(carrier_info_dict)
-        self._tech_data_source_dict.union(base_tech_data, allow_override=True)
-
-        tech_dict = AttrDict({k: {} for k in techs.values})
-        tech_dict.union(self.model_definition["techs"], allow_override=True)
-        self.model_definition["techs"] = tech_dict
-
-    def _update_node_def_from_data_source(self, data_source: xr.Dataset) -> None:
-        """Create a dummy node definition dictionary from the dimensions defined across all data sources.
-
-        In addition, if transmission technologies are defined then update `to`/`from` params in the technology definition dictionary.
-
-        This definition dictionary will ensure that the minimal YAML content is still possible.
-
-        This function should be run _after_ `self._update_tech_def_from_data_source`.
-
-        Args:
-            data_source (xr.Dataset): Data loaded from one file.
-
-        """
-        nodes = data_source.nodes
-        techs_incl_inheritance = self._inherit_defs("techs")
-        node_tech_vars = data_source[
-            [
-                k
-                for k, v in data_source.data_vars.items()
-                if "nodes" in v.dims and "techs" in v.dims
-            ]
-        ]
-        other_dims = [i for i in node_tech_vars.dims if i not in ["nodes", "techs"]]
-
-        is_defined = (
-            node_tech_vars.fillna(False).any(other_dims).to_dataframe().any(axis=1)
-        )
-        node_tech_dict = AttrDict({i: {"techs": {}} for i in nodes.values})
-        for node, tech_dict in node_tech_dict.items():
-            try:
-                techs_this_node = is_defined[is_defined].xs(node, level="nodes").index
-            except KeyError:
-                continue
-            for tech in techs_this_node:
-                if techs_incl_inheritance[tech].get("parent", None) == "transmission":
-                    if "from" in self._tech_data_source_dict[tech]:
-                        self._tech_data_source_dict.set_key(f"{tech}.to", node)
-                    else:
-                        self._tech_data_source_dict.set_key(f"{tech}.from", node)
-                else:
-                    tech_dict["techs"][tech] = None
-
-        node_tech_dict.union(
-            self.model_definition.get("nodes", AttrDict()), allow_override=True
-        )
-        self.model_definition["nodes"] = node_tech_dict
+            self.model_data[var_name] = var_data.assign_attrs(
+                is_result=False, **self.param_attrs.get(var_name, {})
+            )
 
     def _get_relevant_node_refs(self, techs_dict: AttrDict, node: str) -> list[str]:
         """Get all references to parameters made in technologies at nodes.
@@ -598,9 +494,7 @@ class ModelDataFactory:
             param_da = param_series.to_xarray()
         else:
             param_da = xr.DataArray(param_data["data"])
-        param_da = param_da.rename(param_name).assign_attrs(
-            self.param_attrs.get(param_name, {})
-        )
+        param_da = param_da.rename(param_name)
         return param_da
 
     def _definition_dict_to_ds(
@@ -721,13 +615,9 @@ class ModelDataFactory:
                 item_base_def.union(item_def, allow_override=True)
             else:
                 item_base_def = item_def
-            updated_item_def, inheritance = self._climb_inheritance_tree(
+            updated_item_def, inheritance = self.climb_inheritance_tree(
                 item_base_def, dim_name, item_name
             )
-            if dim_name == "techs" and item_name in self._tech_data_source_dict:
-                _data_source_dict = deepcopy(self._tech_data_source_dict[item_name])
-                _data_source_dict.union(updated_item_def, allow_override=True)
-                updated_item_def = _data_source_dict
 
             if not updated_item_def.get("active", True):
                 LOGGER.debug(
@@ -743,58 +633,6 @@ class ModelDataFactory:
             updated_defs[item_name] = updated_item_def
 
         return updated_defs
-
-    def _climb_inheritance_tree(
-        self,
-        dim_item_dict: AttrDict,
-        dim_name: Literal["nodes", "techs"],
-        item_name: str,
-        inheritance: Optional[list] = None,
-    ) -> tuple[AttrDict, Optional[list]]:
-        """Follow the `inherit` references from `nodes` to `node_groups` / from `techs` to `tech_groups`.
-
-        Abstract group definitions (those in `node_groups`/`tech_groups`) can inherit each other, but `nodes`/`techs` cannot.
-
-        This function will be called recursively until a definition dictionary without `inherit` is reached.
-
-        Args:
-            dim_item_dict (AttrDict):
-                Dictionary (possibly) containing `inherit`. If it doesn't contain `inherit`, the climbing stops here.
-            dim_name (Literal[nodes, techs]):
-                The name of the dimension we're working with, so that we can access the correct `_groups` definitions.
-            item_name (str):
-                The current position in the inheritance tree.
-            inheritance (Optional[list], optional):
-                A list of items that have been inherited (starting with the oldest).
-                If the first `dim_item_dict` does not contain `inherit`, this will remain as None.
-                Defaults to None.
-
-        Raises:
-            KeyError: Must inherit from a named group item in `node_groups` (for `nodes`) and `tech_groups` (for `techs`)
-
-        Returns:
-            tuple[AttrDict, Optional[list]]: Definition dictionary with inherited data and a list of the inheritance tree climbed to get there.
-        """
-        dim_name_singular = dim_name.removesuffix("s")
-        dim_group_def = self.model_definition.get(f"{dim_name_singular}_groups", None)
-        to_inherit = dim_item_dict.get("inherit", None)
-        if to_inherit is None:
-            updated_dim_item_dict = dim_item_dict
-        elif dim_group_def is None or to_inherit not in dim_group_def:
-            raise KeyError(
-                f"({dim_name}, {item_name}) | Cannot find `{to_inherit}` in inheritance tree."
-            )
-        else:
-            base_def_dict, inheritance = self._climb_inheritance_tree(
-                dim_group_def[to_inherit], dim_name, to_inherit, inheritance
-            )
-            updated_dim_item_dict = deepcopy(base_def_dict)
-            updated_dim_item_dict.union(dim_item_dict, allow_override=True)
-            if inheritance is not None:
-                inheritance.append(to_inherit)
-            else:
-                inheritance = [to_inherit]
-        return updated_dim_item_dict, inheritance
 
     def _deactivate_item(self, **item_ref):
         for dim_name, item_name in item_ref.items():
@@ -915,91 +753,6 @@ class ModelDataFactory:
                         f"(parameters, {param_name}) | Adding a new value to the "
                         f"`{coord_name}` model coordinate: {new_coord_data.values}"
                     )
-
-    @staticmethod
-    def _listify_if_defined(source: DataSource, key: str) -> Optional[list]:
-        vals = source.get(key, None)
-        if vals is not None:
-            vals = listify(vals)
-        return vals
-
-    @staticmethod
-    def _carrier_info_dict_from_model_data_var(data_source: xr.Dataset) -> AttrDict:
-        """Convert `carrier_in`/`carrier_out` loaded from a data source into YAML-esque format.
-
-        Only used when either variable is defined in a data source.
-        Conversion from e.g.,
-
-        ```
-        nodes | techs | carriers | carrier_in
-        foo   | bar   | baz      | 1
-        ```
-
-        to
-        ```yaml
-        techs:
-          bar:
-            carrier_in: baz
-        ```
-
-        !!! note
-            This will drop the `nodes` dimension entirely as we expect these parameters to be defined at the tech level in YAML.
-
-        Args:
-            data_source (xr.Dataset): data loaded from file that may contain `carrier_in`/`carrier_out`.
-
-        Returns:
-            AttrDict:
-                If present in `data_source`, a valid Calliope YAML format for `carrier_in`/`carrier_out`.
-                If not, an empty dictionary.
-        """
-
-        def __extract_carriers(grouped_series):
-            carriers = list(
-                set(
-                    grouped_series.dropna(
-                        subset=[f"carrier_{direction}"]
-                    ).carriers.values
-                )
-            )
-            if len(carriers) == 1:
-                carriers = carriers[0]
-            if carriers:
-                return {f"carrier_{direction}": carriers}
-            else:
-                return np.nan
-
-        carrier_info_dict = AttrDict()
-        for direction in ["in", "out"]:
-            if f"carrier_{direction}" not in data_source:
-                continue
-            carrier_info_dict.union(
-                AttrDict(
-                    data_source[f"carrier_{direction}"]
-                    .to_series()
-                    .reset_index("carriers")
-                    .groupby("techs")
-                    .apply(__extract_carriers)
-                    .dropna()
-                    .to_dict()
-                )
-            )
-
-        return carrier_info_dict
-
-    @staticmethod
-    def _compare_axis_names(
-        loaded_names: list, defined_names: list, axis: str, filename: str
-    ):
-        if any(
-            name is not None
-            and not isinstance(name, int)
-            and name != defined_names[idx]
-            for name, idx in enumerate(loaded_names)
-        ):
-            raise ValueError(
-                f"(data_sources, {filename}) | Trying to set names for {axis} but names in the file do no match names provided | in file: {loaded_names} | defined: {defined_names}"
-            )
 
     @staticmethod
     def _update_one_way_links(node_from_data: dict, node_to_data: dict):
